Reorganise repo
All solvitaire content has been shoved into a subdirectory to tidy up
the root of the project. This also means solvitaire is treated similarly
to other dependencies of the project, and allows other dependencies to
be added easily.

There's also a lot of other work done here, such as usign a thread pool
library to manage the running of the voters.