Reveal stock cards if stock redeals allowed
If stock redeals are allowed, then the player can go through the stock
once at the start of the game and reset it, bringing the game back to
the original game state. This allows the player to see all cards in the
stock, even if the stock isn't explicitly face-up.

This can be used to reduce the number of voters run, since the stock
will not be hidden and fewer reveals means fewer full resets. As such,
this optimisation seems to reduce solving time by roughly half.

This optimisation also theoretically increases solvability percentages
in affected games, since the system has fewer face-down cards to
generate misleading permutations with.