Clean up rule description files
The files for describing rules have changed a bit: they are no longer
`.json` files (although they are still parsed by a JSON parser). The
description files are now their own format, with a little less
boilerplate required than for JSON files.

As a result, rule description files can be given shebangs and be made
executable, allowing a user to run the rule file instead of `centre`
directly. The rules files in the repository are set up to run `centre`
if it is on the user's path.