Skip to 'solved' when there are no hidden cards
If a game-state can be solved by a voter from a position where there are
no hidden cards, the game can be solved by metasol also. The rest of the
game can be skipped, which saves time.

If the sequence of moves taken is required, the remaining moves can be
taken from a successful voter.