Change threadpool library
thpool was great as a simple threadpool library, but lacked features
required in the future of the project.

CTPL is a C++ library, and uses futures. This takes the project a littel
further from becoming a pure C library, but that can be fixed once it
doesn't contribute to my grade.